+++ src/core/Track.cpp
@@ -1,7 +1,7 @@
 #include "Track.hpp"
 #include "core.hpp"
 
-Game::Track Game::getNthTrack(unsigned short num, float start, float length) {
+Game::Track Game::getNthTrack(unsigned num, float start, float length) {
 	using Game::DIRSEP;
 	std::stringstream ss;
 	ss << Game::pwd << DIRSEP << "assets" << DIRSEP << "music" << DIRSEP << "music" << num << ".ogg";
