+++ src/core/components/Scored.cpp
@@ -3,7 +3,7 @@
 
 using Game::Scored;
 
-void Scored::setTarget(unsigned short t) {
+void Scored::setTarget(unsigned t) {
 	if (target != -1) return; // can only set target once
 	if (0 >= t || t > Game::MAX_PLAYERS)
 		throw std::logic_error("setTarget set to a number greater than MAX_PLAYERS or < 1!");
