+++ src/lifish/controls.hpp
@@ -10,7 +10,7 @@ namespace Game {
 
 namespace Controls {
 
-enum Control : unsigned short {
+enum Control : unsigned {
 	CTRL_UP    = 0,
 	CTRL_DOWN  = 1,
 	CTRL_LEFT  = 2,
@@ -18,12 +18,12 @@ enum Control : unsigned short {
 	CTRL_BOMB  = 4
 };
 
-constexpr static unsigned short CONTROLS_NUM = 5;
+constexpr static unsigned CONTROLS_NUM = 5;
 
 extern std::array<sf::Keyboard::Key, CONTROLS_NUM> players[Game::MAX_PLAYERS];
-extern std::array<unsigned int, Game::MAX_PLAYERS> joystickBombKey;
+extern std::array<unsigned, Game::MAX_PLAYERS> joystickBombKey;
 /** -1 if i-th player doesn't use joystick, else the joystick index. */
-extern std::array<short, Game::MAX_PLAYERS> useJoystick;
+extern std::array<int, Game::MAX_PLAYERS> useJoystick;
 
 inline Control fromString(const std::string& name) {
 	if (name == "up" || name == "UP") return CTRL_UP;
