+++ src/lifish/entities/Bomb.hpp
@@ -18,7 +18,7 @@ class Killable;
 class Bomb : public Game::Entity {
 	Game::Clock *fuseClock = nullptr;
 	sf::Time fuseTime;
-	unsigned short radius;
+	unsigned radius;
 	
 	bool ignited = false;
 	bool exploded = false;
@@ -35,7 +35,7 @@ public:
 
 	explicit Bomb(const sf::Vector2f& pos, const Game::Player& source, 
 			const sf::Time& fuseTime = Game::Conf::Bomb::DEFAULT_FUSE, 
-			const unsigned short radius = Game::Conf::Bomb::DEFAULT_RADIUS);
+			const unsigned radius = Game::Conf::Bomb::DEFAULT_RADIUS);
 
 	void update() override;
 	
@@ -44,8 +44,8 @@ public:
 	/** Manually set this bomb to explode after 50 ms */
 	void ignite();
 
-	unsigned short getRadius() const { return radius; }
-	void setRadius(unsigned short r) { radius = r; }
+	unsigned getRadius() const { return radius; }
+	void setRadius(unsigned r) { radius = r; }
 
 	/** Returns true if this bomb's fuse is over and the bomb should blow off. */
 	bool isFuseOver() const;
